==> 04_valorizacion_con_colateral.ipynb <==
# Valorización con Colateral

**Ametrano, Ferdinando M. and Bianchetti, Marco**, Everything You Always Wanted to Know About Multiple Interest Rate Curve Bootstrapping but Were Afraid to Ask (April 2, 2013). Available at SSRN: https://ssrn.com/abstract=2219548 or http://dx.doi.org/10.2139/ssrn.2219548

Se puede suponer que las cantidades captadas o colocadas por una contraparte que participa en el mercado OTC están asociadas a una cuenta de financiamiento (*funding account*) con el valor $B_{\alpha}\left(t\right)$ en $t$. El índice $\alpha$ denota la fuente específica de financiamiento.

Se asume la siguiente dinámica:

$$
\begin{equation}
dB_{\alpha}\left(t\right)=r_{\alpha}\left(t\right)B_{\alpha}\left(t\right)dt
\end{equation}
$$


$$
\begin{equation}
B_{\alpha}\left(0\right)=1
\end{equation}
$$


$$
\begin{equation}
B_{\alpha}\left(t\right)=\exp\int_{0}^{t}r_{\alpha}\left(u\right)du
\end{equation}
$$

Hemos ya identificado dos fuentes de financiamiento para el mercado OTC:

- La fuente de financiamiento general (o de tesorería), denotada con $B_f$ asociada a la operativa usual en mercados monetarios y de bonos. Podemos pensar que, en este caso, $r_f=Libor+spread$


- La cuenta de colateral, $B_c$, asociada a un CSA para la cual típicamente $r_c=ON$.

## Definición

Un CSA perfecto es un CSA con las siguientes características:

- margen inicial igual a 0
- totalmente simétrico (las condiciones aplican a ambas contrapartes por igual)
- colateral sólo en *cash*
- threshold igual a 0
- MTA igual a 0
- periodicidad de cálculo continua
- tasa de colateral $r_c\left(t\right)$
- traspaso instantáneo de colateral
- sin reinversión de colateral

Como consecuencia se tiene que:

$$
\begin{equation}
B_c\left(t\right)=\Pi\left(t\right)
\end{equation}
$$

donde $\Pi\left(t\right)$ es el valor de la cartera de derivados asociada al CSA.

## Teorema (valorización con colateral)

Sea $\Pi$ un instrumento derivado que vence en $T$ escrito sobre el activo $X$ que obedece la siguiente ecuación diferencial estocástica:


$$
\begin{equation}
dX\left(t\right)=\mu^P\left(t,X\right)X\left(t\right)dt+\sigma\left(t,X\right)X\left(t\right)dW^P\left(t\right),
\end{equation}
$$


$$
\begin{equation}
X\left(0\right)=X_0
\end{equation}
$$

donde $t\in\mathbb{R}_+$, $X\in\mathbb{R}$, $\mu^P\in\mathbb{R}_+\times\mathbb{R}\rightarrow\mathbb{R}$, $\sigma\in\mathbb{R}_+\times\mathbb{R}\rightarrow\mathbb{R}_+$, $W^P\in\mathbb{R}$ es un movimiento browniano de dimension 1 sobre $\left(\Omega,\mathscr{F},P\right)$ y $P$ es la medida histórica (u objetiva). Asumiendo colateral perfecto, el precio del derivado $\Pi\left(t\right)$ en $t\lt T$ satisface:

$$
\begin{equation}
\overset{\wedge}{\mathcal{D}}_{rf}\Pi\left(t,X\right)=r_c\left(t\right)\Pi\left(t,X\right)
\end{equation}
$$


$$
\begin{equation}
\overset{\wedge}{\mathcal{D}}_{rf}=\frac{\partial}{\partial t}+r_f\left(t\right)X\left(t\right)\frac{\partial}{\partial X}+\frac{1}{2}\sigma^2\left(t\right)X^2\left(t\right)\frac{\partial^2}{\partial X^2}
\end{equation}
$$


y está dado también por:

$$
\begin{equation}
\Pi\left(t,X\right)=\mathbb{E}_t^{Q_f}\left[D_c\left(t,T\right)\Pi\left(T,X\right)\right]
\end{equation}
$$

$$
\begin{equation}
D_c\left(t,T\right)=\exp\left[-\int_t^Tr_c\left(u\right)du\right]
\end{equation}
$$


donde $Q_f$ es la medida de probabilidad asociada a la cuenta de financiamiento $B_f$ de modo que:


$$
\begin{equation}
dX\left(t\right)=r_f\left(t\right)X\left(t\right)dt+\sigma\left(t,X\right)X\left(t\right)dW^{Q_f}\left(t\right),
\end{equation}
$$

### Corolario (en la medida forward)

Vale la expresión:

$$
\begin{equation}
\Pi\left(t,X\right)=P_c\left(t,T\right)\mathbb{E}_{t}^{Q_f^T}\left[\Pi\left(T,X\right)\right]
\end{equation}
$$

$$
\begin{equation}
P_c\left(t,T\right)=\mathbb{E}^{Q_f}\left[D_c\left(t,T\right)\right]
\end{equation}
$$

donde $Q_f^T$ es la probabilidad asociada a $P_c\left(t,T\right)$ (la medida $T$-forward).
