==> 06_construccion_curva_libor_usd_3m.ipynb <==
# Construcción Curva Libor USD 3M

from IPython.display import Image

## Valor de un Cupón Libor

Tenemos que (notebook 4):

$$
\begin{equation}
\Pi\left(t,X\right)=\mathbb{E}_t^{Q_f}\left[D_c\left(t,T\right)\Pi\left(T,X\left(T\right)\right)\right]
\end{equation}
$$

$$
\begin{equation}
D_c\left(t,T\right)=\exp\left[-\int_t^Tr_c\left(u\right)du\right]
\end{equation}
$$

$$
\begin{equation}
\mathbb{E}_t^{Q_f}\left[D_c\left(t,T\right)\right]=P_c\left(t,T\right)
\end{equation}
$$

$$
\begin{equation}
D_c\left(t,T\right)={B_c\left(t,T\right)}^{-1}
\end{equation}
$$

También vamos a necesitar:

$$
\begin{equation}
\Pi\left(t,X\right)=P_c\left(t,T\right)\mathbb{E}_{t}^{Q_f^T}\left[\Pi\left(T,X\right)\right]
\end{equation}
$$

$$
\begin{equation}
P_c\left(t,T\right)=\mathbb{E}^{Q_f}\left[D_c\left(t,T\right)\right]
\end{equation}
$$

donde $Q_f^T$ es la probabilidad asociada a $P_c\left(t,T\right)$ (la medida $T$-forward).

Sea $L\left(T\right)$ el valor de la Libor USD de $tenor=3M$ al tiempo $T$ y sea $\Delta\left(T,T_{3M}\right)$ la fracción de año asociada a esa Libor y $T_{3M}$ es la fecha $3M$ después de $T$. Para un cupón flotante a Libor (suponiendo un nocional igual a 1), tenemos que:

$$
\begin{equation}
\Pi\left(T,X\left(T\right)\right)=L\left(T\right)\cdot\Delta\left(T,T_{3M}\right)
\end{equation}
$$

Y por lo tanto,

$$
\begin{equation}
\Pi\left(t,X\left(t\right)\right)=\mathbb{E}_t^{Q_f}\left[D_c\left(t,T_{3M}\right)L\left(T\right)\cdot\Delta\left(T,T_{3M}\right)\right]
\end{equation}
$$

Finalmente, utilizando (5) y (6) y la linealidad del valor esperado se llega a:

$$
\begin{equation}
\Pi\left(t,X\left(t\right)\right)=\Delta\left(T,T_{3M}\right)\cdot P_c\left(t,T\right)\cdot\mathbb{E}_t^{Q_f^T}\left[L\left(T\right)\right]
\end{equation}
$$

## Cotizaciones Libor USD 3M vs Tasa Fija

Image(url="img/libor_usd_3m_fixed_swap_quotes.gif", width=900, height=720)

Como se puede observar en la imagen anterior, existen cotizaciones de swaps Libor USD 3M versus tasa fija para plazos hasta 30Y. Utilizando la fórmula (9) y lo que ya vimos sobre la construcción de una curva a partir de las cotizaciones de OIS ¿Cómo se escribe la fórmula de valorización para un swap Libor?

Para fijar las ideas, consideremos el swap cuyo $tenor=1Y$. Recordemos (notebook 2, 2.2.1) que, en estos swaps, la pata fija tiene periodicidad semestral y convención de tasa 30/360, mientras que la pata flotante tiene trimestral y convención de tasa Act/360.

Tenemos que, el valor de la pata fija es:

$$
\begin{equation}
\Pi^{fija,1Y}\left(t\right)=P_c\left(t,T_1\right)\cdot r\cdot\delta\left(T_0,T_1\right)+P_c\left(t,T_2\right)\left(1+r \cdot \delta\left(T_1,T_2\right)\right)
\end{equation}
$$

Donde $P_c\left(t,T\right)$ proviene de la curva OIS (ver notebook 5) y $\delta\left(T_i,T_j\right)$ representa la fracción de año en convención 30/360 entre $T_i$ y $T_j$.

Por otro lado, el valor de la pata flotante es:

$$
\begin{equation}
\Pi^{flot,1Y}\left(T_0\right)=\sum\limits_{j=1}^{3}\left\{P_c\left(T_0,T_{3M}^j\right)\cdot\mathbb{E}_t^{Q_f^T}\left[L\left(T_j\right)\right]\cdot\Delta\left(T_j,T_{3M}^j\right)\right\}+P_c\left(T_0,T_{4}\right)\cdot\left(1+\cdot\mathbb{E}_t^{Q_f^T}\left[L\left(T_4\right)\right]\cdot\Delta\left(T_4,T_{3M}^4\right)\right)
\end{equation}
$$

Donde $\Delta\left(T_i,T_{3M}^i\right)$ representa la fracción de año en convención Act/360 correspondiente la Libor que parte en $T_i$.

Podemos observar que esta ecuación tiene 4 incógnitas, y por lo tanto, para poder aplicar un procedimiento de bootstrapping como para los OIS, se requiere disponer de data de corto plazo, que permita deducir algunos valores iniciales.

Afortunadamente, dicha data sí existe y se puede apreciar en la imagen siguiente, lo que hace inncesario resolver una ecuación para este primer dato.

Sin embargo, para el swap de 3Y, si se debe resolver una ecuación.

**TODO:** mejorar esta redacción.

Image(url="img/20201016_libor_curve.gif", width=900, height=720)

La primera de las tasas marcadas en rojo corresponde a la Libor USD 3M del 16-10-2020. Las siguientes 6 tasas, se obtienen de las cotizaciones de futuros [eurodollar](https://www.cmegroup.com/trading/interest-rates/stir/eurodollar.html) con esos *nemotécnicos*.
