==> 08_tarea_1.ipynb <==
# Tarea 1

Implementar la clase `OisCashflow`.

## Configuración Inicial

from finrisk import QC_Financial_3 as Qcf
from modules import auxiliary as aux
from dataclasses import dataclass
from enum import Enum
import pandas as pd

class BusCal(Enum):
    NY = 1
    SCL = 2

def get_cal(code: BusCal) -> Qcf.BusinessCalendar:
    """
    """
    if code == BusCal.NY:
        cal = Qcf.BusinessCalendar(Qcf.QCDate(1, 1, 2020), 20)
        for agno in range(2019, 2071):
            f = Qcf.QCDate(12, 10, agno)
            if f.week_day() == Qcf.WeekDay.SAT:
                cal.add_holiday(Qcf.QCDate(14, 10, agno))
            elif f.week_day() == Qcf.WeekDay.SUN:
                cal.add_holiday(Qcf.QCDate(13, 10, agno))
            elif f.week_day() == Qcf.WeekDay.MON:
                cal.add_holiday(Qcf.QCDate(12, 10, agno))
            elif f.week_day() == Qcf.WeekDay.TUE:
                cal.add_holiday(Qcf.QCDate(11, 10, agno))
            elif f.week_day() == Qcf.WeekDay.WED:
                cal.add_holiday(Qcf.QCDate(10, 10, agno))
            elif f.week_day() == Qcf.WeekDay.THU:
                cal.add_holiday(Qcf.QCDate(9, 10, agno))
            else:
                cal.add_holiday(Qcf.QCDate(8, 10, agno))
        cal.add_holiday(Qcf.QCDate(11, 11, 2019))
        cal.add_holiday(Qcf.QCDate(28, 11, 2019))
        cal.add_holiday(Qcf.QCDate(25, 12, 2019))
        cal.add_holiday(Qcf.QCDate(20, 1, 2020))
        cal.add_holiday(Qcf.QCDate(17, 2, 2020))
        cal.add_holiday(Qcf.QCDate(10, 4, 2020))
        cal.add_holiday(Qcf.QCDate(25, 5, 2020))
        cal.add_holiday(Qcf.QCDate(3, 7, 2020))
        cal.add_holiday(Qcf.QCDate(7, 9, 2020))
        # cal.add_holiday(Qcf.QCDate(30, 9, 2020)) Este no corresponde y se detectó en los tests.
        cal.add_holiday(Qcf.QCDate(15, 2, 2021))
        
    return cal

## `Qcf.time_series`

ts = Qcf.time_series()

fecha1 = Qcf.QCDate(13, 1, 1969)
ts[fecha1] = 19690113

fecha2 = Qcf.QCDate(14, 1, 1969)
ts[fecha2] = 19690114

print(f'ts[fecha1]: {ts[fecha1]}\n')

dl = Qcf.time_series_dates(ts)
for d in dl:
    print(d)
print()
    
vl = Qcf.time_series_values(ts)
for v in vl:
    print(v)

## Clase `OisCashflow`

La clase `OisCashflow` es una [`dataclass`](https://realpython.com/python-data-classes/). La tarea consiste en:

- Implementar los 3 métodos de la clase que están con `pass`.
- Implementar la función `present_value`.
- Implementar la función `set_expected_rate`.

@dataclass  # syntactic sugar
class OisCashflow:
    start_date: Qcf.QCDate
    end_date: Qcf.QCDate
    settlement_date: Qcf.QCDate
    notional: float
    currency: Qcf.QCCurrency
    amortization: float
    amort_is_cashflow: bool
    interest_rate: Qcf.QCInterestRate
    on_index: Qcf.InterestRateIndex
    spread: float
    gearing: float
    calendar: Qcf.BusinessCalendar  # tuve que agregar esta variable

    def get_accrued_rate(self, accrual_date: Qcf.QCDate, fixings: Qcf.time_series) -> float:
        """
        Calcula la tasa equivalente desde `self.start_date` a `accrual_date`. La tasa equivalente
        se calcula como:

        (P - 1) * 360 / (accrual_date - self.start_date)

        donde P es el producto de los factores de capitalización de todas las tasas overnight
        entre `start_date` y `accrual_date`. Los valores de esas tasas deben estar almacenados
        en la variable `fixings`.

        Ver la documentación de `QC_Financial_3` para el uso y funcionamiento de los objetos
        de tipo Qcf.time_series.
        """
        if self.calendar.next_busy_day(accrual_date) > accrual_date:
            raise ValueError("Acrrual date must be a working day.")
        fecha = self.start_date
        if accrual_date <= fecha:
            return 0.0
        factor = 1.0
        while fecha < accrual_date:
            fecha_siguiente = self.calendar.shift(fecha, 1)
            tasa = self.on_index.get_rate()
            try:
                tasa.set_value(fixings[fecha])
            except Exception as e:
                raise ValueError(
                    f'Se produjo el error: {str(e)}. Fecha de cálculo: {fecha.description(False)}')
            factor *= tasa.wf(fecha, fecha_siguiente)
            fecha = fecha_siguiente
        return self.interest_rate.get_rate_from_wf(factor, self.start_date, accrual_date)

    def get_accrued_interest(self, accrual_date: Qcf.QCDate, fixings: Qcf.time_series) -> float:
        """
        Calcula el interés (plata) devengado desde `self.start_date` a `accrual_date` utilizando la tasa equivalente
        que se calcula con el método anterior.

        Los valores de las  tasas overnight deben estar almacenados en la variable `fixings`.

        Ver la documentación de `QC_Financial_3` para el uso y funcionamiento de los objetos
        de tipo Qcf.time_series.
        """
        valor_tasa = self.get_accrued_rate(accrual_date, fixings)
        self.interest_rate.set_value(valor_tasa)
        return self.notional * (self.interest_rate.wf(self.start_date, accrual_date) - 1.0)

    def amount(self, fixings: Qcf.time_series = None) -> float:
        """
        Calcula el flujo total al vencimiento (amortización más intereses devengados hasta self.end_date).

        Los valores de las  tasas overnight deben estar almacenados en la variable `fixings`.

        Ver la documentación de `QC_Financial_3` para el uso y funcionamiento de los objetos
        de tipo Qcf.time_series.
        """
        if fixings is None:
            return self.notional * self.interest_rate.wf(self.start_date, self.end_date)
        else:
            return self.get_accrued_interest(self.end_date, fixings) + self.notional

Construcción de un objeto `Qcf.InterestRateIndex`.

codigo = 'SOFR'
tasa_on = Qcf.QCInterestRate(.0, Qcf.QCAct360(), Qcf.QCLinearWf())
fixing_lag = Qcf.Tenor('0d')
tenor = Qcf.Tenor('1d')
fixing_calendar = get_cal(BusCal.NY)     
settlement_calendar = fixing_calendar  
sofr = Qcf.InterestRateIndex(
    codigo,
    tasa_on,
    fixing_lag,
    tenor,
    fixing_calendar,
    settlement_calendar,
    Qcf.QCUSD()
)

Construcción de una instancia de `OisCashflow`.

ois = OisCashflow(
    Qcf.QCDate(1, 10, 2019),
    Qcf.QCDate(1, 10, 2020),
    Qcf.QCDate(1, 10, 2020),
    10000000,
    Qcf.QCUSD(),
    1000000,
    True,
    Qcf.QCInterestRate(0.0, Qcf.QCAct360(), Qcf.QCLinearWf()),
    sofr,
    0,
    1,
    settlement_calendar
)

## Funciones

### `set_expected_rate(...)`

Las funciones que hacen este trabajo en `QC_Finacial_3` no retornan ningún valor, pero cambian el estado del cashflow, al dejar almacenado dentro de él el valor de la tasa flotante (o el valor de los índices ICP). En la clase `OisCashflow` este cambio de estado puede lograrse cambiando el valor de la tasa en el objeto `interest_rate`.

def set_expected_rate(
        val_date: Qcf.QCDate,
        ois_cashflow: OisCashflow,
        fixings: Qcf.time_series,
        zcc: Qcf.ZeroCouponCurve) -> None:
    """
    Esta función opera de la misma forma que la análoga función de `QC_Financial_3`.
    Ver por ejemplo los casos cuando usamos el objeto Qcf.ForwardRates() en el notebook 9.

    Para un OIS calculando en t0 un cashflow entre (t0 <) t1 < t2
    ICP(t0) * wf(to, t2) = ICP(t2), ICP(t0) * w(t0, t1) = ICP(t1) -> ICP(t2) / ICP(t1) = ICP(t1,t2)
    tasa(t1, t2) = [ICP(t1, t2) - 1] * 360 / (t2-t1)
    El ICP se puede pensar como el account asociado a la tasa interbancaria chilena, que es el índice
    overnight chileno.

    Cuando t1 < t0 < t2 tener un poquito de ojo ... porque el cashflow ya tiene fixings.

    Se distinguen 3 casos (en todos ellos `zcc.wf()` significa `1 / zcc.get_discount_factor_at()`:

    - `val_date <= ois_cashflow.start_date`: se calculan los dos plazos p1, p2 desde `val_date` hasta
    `ois_cashflow.start_date` y `ois_cashflow.end_date` respectivamente. Se define `account(val_date) = 1.0` y luego,
      - `account(ois_cashflow.start_date) = 1.0 * zcc.wf(p1) := wf1`
      - `account(ois_cashflow.end_date) = 1.0 * zcc.wf(p2) := wf2`
      - `tasa_esperada = (wf2 / wf2 - 1.0) * 360.0 / (p2 - p1)`
      - Finalmente, `ois_cashflow.interest_rate.set_value(tasa_esperada)`.

    - `ois_cashflow.start_date < val_date < ois_cashflow.end_date`:
      - se calcula el plazo `p1` desde `ois_cashflow.end_date` hasta `val_date`.
      - Se calcula `wf1 = 1 + ois_cashflow.get_accrued_interest(val_date, fixings) / ois_cashflow.notional`.
      - se calcula el plazo `p2` desde `val_date` hasta `ois_cashflow.end_date`.
      - Se calcula `wf2 = zcc.wf(p2)`.
      - `tasa_esperada = (wf2 / wf2 - 1.0) * 360.0 / (p1 + p1)`.
      - Finalmente, `ois_cashflow.interest_rate.set_value(tasa_esperada)`.

    - `ois_cashflow.end_date <= val_date`: en este caso el flujo debe ser conocido, y por lo tanto:
      - `tasa_esperada = ois_cashflow.accrued_rate(ois_cashflow.end_date)`.
      - Finalmente, `ois_cashflow.interest_rate.set_value(tasa_esperada)`.
    """
    if val_date < ois_cashflow.start_date:
        p1 = val_date.day_diff(ois_cashflow.start_date)
        p2 = val_date.day_diff(ois_cashflow.end_date)
        wf1 = 1 / zcc.get_discount_factor_at(p1)
        wf2 = 1 / zcc.get_discount_factor_at(p2)
        tasa_esperada = (wf2 / wf1 - 1) * 360 / (p2 - p1)
    elif ois_cashflow.start_date <= val_date < ois_cashflow.end_date:
        p1 = ois_cashflow.start_date.day_diff(val_date)
        wf1 = 1 + ois_cashflow.get_accrued_interest(
            val_date, fixings) / ois_cashflow.notional
        p2 = val_date.day_diff(ois_cashflow.end_date)
        wf2 = 1 / zcc.get_discount_factor_at(p2)
        tasa_esperada = (wf1 * wf2 - 1) * 360 / (p1 + p2)
    else:
        tasa_esperada = ois_cashflow.get_accrued_rate(
            ois_cashflow.end_date, fixings)
    ois_cashflow.interest_rate.set_value(tasa_esperada)
    return

### `present_value(...)`

Para que esta función opere de la misma forma que los métodos de la clase `PresentValue` de `QC_Fimancial_3` se requiere tener un método `amount` en `OisCashflow` que no requiera argumentos y retorne el flujo final utilizando el valor de la tasa que se fijó con la función `set_expected_rate`.

En Python no se puede sobrecargar una función o método de una clase (mismo nombre para una función que tiene más de una firma). Sin embargo, se puede lograr un efecto similar transformando en opcional el argumento de la versión ya implementada de `amount`.

def present_value(
        val_date: Qcf.QCDate,
        ois_cashflow: OisCashflow,
        zcc: Qcf.ZeroCouponCurve) -> float:
    """
    Esta función opera de la misma forma que la análoga función de `QC_Financial_3`.
    Para probar esta función utilizar la data de que está en data/20201012_built_sofr_zero.xlsx.
    Para traer a valor presente debes respetar la convención de las tasas de la curva.

    NOTA: la variable ois_cashflow debe haber 'pasado' por la función set_expected_rate para que
    tenga los intereses de los fixings futuros.
    """
    plazo = val_date.day_diff(ois_cashflow.settlement_date)
    if plazo <= 0:
        return 0.0
    else:
        return ois_cashflow.amount() * zcc.get_discount_factor_at(plazo)

## Tests

Construcción de una instancia de `OisCashflow`.

ois = OisCashflow(
    Qcf.QCDate(1, 10, 2019),
    Qcf.QCDate(1, 10, 2020),
    Qcf.QCDate(1, 10, 2020),
    10000000,
    Qcf.QCUSD(),
    1000000,
    True,
    Qcf.QCInterestRate(0.0, Qcf.QCAct360(), Qcf.QCLinearWf()),
    sofr,
    0,
    1,
    settlement_calendar
)

### Objeto `fixings`

df_fixings = pd.read_excel('data/SOFR-10012019-10292020.xls', sheet_name='nice format')
df_fixings.columns = ['fecha', 'nombre', 'valor']
df_fixings['valor'] /= 100

df_fixings.head().style.format({'valor': '{:.4%}'})

fixings = Qcf.time_series()
for row in df_fixings.itertuples():
    fixings[Qcf.build_qcdate_from_string(row.fecha)] = row.valor

fixings[Qcf.QCDate(13, 1, 2020)]

### `OisCashflow.accrued_rate`

accrued_rate = ois.get_accrued_rate(Qcf.QCDate(15, 6, 2020), fixings)
print(f'accrued rate: {accrued_rate:.8%}')

Se verificará el cálculo usando `df_fixings`.

df_fixings.head().style.format({'valor': '{:.4%}'})

Se agrega la columna `next_date`, servirá para calcular el factor de capitalización de cada tasa.

def next_date(fecha:str, calendario: Qcf.BusinessCalendar) -> str:
    qfecha = Qcf.build_qcdate_from_string(fecha)
    next_fecha = calendario.shift(qfecha, 1)
    return next_fecha.description(False)

df_fixings['next_fecha'] = df_fixings.apply(
    lambda row: next_date(row['fecha'], settlement_calendar),
    axis=1
)

df_fixings.head().style.format({'valor': '{:.4%}'})

Se calcula ahora el factor de capitalización.

def factor_cap(fecha: str, next_fecha: str, valor: float) -> float:
    qfecha = Qcf.build_qcdate_from_string(fecha)
    qnext_fecha = Qcf.build_qcdate_from_string(next_fecha)
    int_rate = Qcf.QCInterestRate(valor, Qcf.QCAct360(), Qcf.QCLinearWf())
    return int_rate.wf(qfecha, qnext_fecha)

df_fixings['factor_capitalizacion'] = df_fixings.apply(
    lambda row: factor_cap(row['fecha'], row['next_fecha'], row['valor']),
    axis=1
)

df_fixings.head().style.format({'valor': '{:.4%}'})

Veamos donde están en `df_fixings` los valores para `start_date` de `ois` y la fecha 15-6-2020 y calculamos el producto de los factores de capitalización desde esa fecha hasta la última fecha de `df_flujos` (que es el primer registro). Luego, el cociente de los dos factores nos dará el factor entre ambas fechas. Finalmente, con ese factor, se calculará `accrued_rate`.

start_date = ois.start_date.description(False)
df_fixings[df_fixings.fecha == start_date]

factor_largo = df_fixings.iloc[:270]['factor_capitalizacion'].prod()
print(factor_largo)
qstart_date = Qcf.build_qcdate_from_string('2019-10-01')

df_fixings[df_fixings.fecha == '2020-06-15']

factor_corto = df_fixings.iloc[:95]['factor_capitalizacion'].prod()
qaccrued_date = Qcf.build_qcdate_from_string('2020-06-15')

dias = qstart_date.day_diff(qaccrued_date)
check_accrued_rate = (factor_largo / factor_corto - 1.0) * 360.0 / dias
print(f'La diferencia es: {accrued_rate - check_accrued_rate:.8%}')

### `OisCashflow.accrued_interest`

accrued_interest = ois.get_accrued_interest(Qcf.QCDate(15, 6, 2020), fixings)
print(f'accrued interest: {accrued_interest:,.2f}')

check_accrued_interest = ois.notional * accrued_rate * dias / 360.0
print(f'La diferencia es: {accrued_interest - check_accrued_interest:,.6f}')

### `OisCashflow.amount`

amount = ois.amount(fixings)
print(f'amount: {amount:,.2f}')

end_date = ois.end_date.description(False)
print(f'end date: {end_date}')
df_fixings[df_fixings.fecha == end_date]

factor_corto_2 = df_fixings.iloc[:19]['factor_capitalizacion'].prod()
dias_2 = qstart_date.day_diff(ois.end_date)
accrued_rate_amount = (factor_largo / factor_corto_2- 1.0) * 360.0 / dias_2
check_amount = ois.notional * (1 + accrued_rate_amount * dias_2 / 360.0)
print(f'La diferencia es: {amount - check_amount:.6f}')

### `set_expected_rate`

Comenzamos cargando los valores de la curva cero cupón que se utilizará para hacer pruebas de `set_expected_rate` y `present_value`.

df_curva = pd.read_excel('data/20201012_built_sofr_zero.xlsx')

df_curva.head().style.format({'tasa': '{:.4%}', 'df': '{:.6%}'})

Con la data se construye un objeto de tipo `Qcf.ZeroCouponCurve`.

zcc = aux.get_curve_from_dataframe(Qcf.QCAct365(),Qcf.QCCompoundWf(), df_curva)

Comienzan los tests.

print(f'Fecha inicial flujo: {ois.start_date.description(False)}')
print(f'Fecha final flujo: {ois.end_date.description(False)}')

# Check primer caso
val_date = Qcf.QCDate(1, 6, 2019)
set_expected_rate(val_date, ois, fixings, zcc)
print(f'Tasa esperada es: {ois.interest_rate.get_value():.8%}')
p1 = val_date.day_diff(ois.start_date)
p2 = val_date.day_diff(ois.end_date)
df1 = zcc.get_discount_factor_at(p1)
df2 = zcc.get_discount_factor_at(p2)
check = (df1 / df2 - 1) * 360 / (p2 - p1)
print(f'Check tasa esperada es: {check:.8%}')

# Check segundo caso borde
val_date = Qcf.QCDate(1, 10, 2019)
set_expected_rate(val_date, ois, fixings, zcc)
print(f'Tasa esperada es: {ois.interest_rate.get_value():.8%}')
p1 = val_date.day_diff(ois.start_date)
p2 = val_date.day_diff(ois.end_date)
df1 = zcc.get_discount_factor_at(p1)
df2 = zcc.get_discount_factor_at(p2)
check = (df1 / df2 - 1) * 360 / (p2 - p1)
print(f'Check tasa esperada es: {check:.8%}')

# Check segundo caso
val_date = Qcf.QCDate(1, 11, 2019)
set_expected_rate(val_date, ois, fixings, zcc)
print(f'Tasa esperada es: {ois.interest_rate.get_value():.8%}')
p1 = ois.start_date.day_diff(val_date)
p2 = val_date.day_diff(ois.end_date)
wf1 = 1 + ois.get_accrued_interest(val_date, fixings) / ois.notional
wf2 = 1 / zcc.get_discount_factor_at(p2)
check = (wf1 * wf2 - 1) * 360 / (p1 + p2)
print(f'Check tasa esperada es: {check:.8%}')

# Check tercer caso borde
val_date = Qcf.QCDate(1, 10, 2020)
set_expected_rate(val_date, ois, fixings, zcc)
print(f'Tasa esperada es: {ois.interest_rate.get_value():.8%}')
check = ois.get_accrued_rate(val_date, fixings)
print(f'Check tasa esperada es: {check:.8%}')

# Check tercer caso
val_date = Qcf.QCDate(15, 10, 2020)
set_expected_rate(val_date, ois, fixings, zcc)
print(f'Tasa esperada es: {ois.interest_rate.get_value():.8%}')
check = ois.get_accrued_rate(ois.end_date, fixings)
print(f'Check tasa esperada es: {check:.8%}')

### `present_value`

# Check primer caso
val_date = Qcf.QCDate(1, 6, 2019)
set_expected_rate(val_date, ois, fixings, zcc)
pv = present_value(val_date, ois, zcc)
print(f'Valor presente es: {pv:,.4f}')
amount = ois.amount()
print(f'Amount: {amount:,.4f}')
amount = ois.amount()
plazo = val_date.day_diff(ois.settlement_date)
df = zcc.get_discount_factor_at(plazo)
print(f'Df: {df:.6%}')
vp_check = df * amount
print(f'VP check es: {vp_check:,.4f}')

# Check segundo caso borde
val_date = Qcf.QCDate(1, 10, 2019)
set_expected_rate(val_date, ois, fixings, zcc)
pv = present_value(val_date, ois, zcc)
print(f'Valor presente es: {pv:,.4f}')
amount = ois.amount()
print(f'Amount: {amount:,.4f}')
amount = ois.amount()
plazo = val_date.day_diff(ois.settlement_date)
df = zcc.get_discount_factor_at(plazo)
print(f'Df: {df:.6%}')
vp_check = df * amount
print(f'VP check es: {vp_check:,.4f}')

# Check segundo caso
val_date = Qcf.QCDate(1, 11, 2019)
set_expected_rate(val_date, ois, fixings, zcc)
pv = present_value(val_date, ois, zcc)
print(f'Valor presente es: {pv:,.4f}')
amount = ois.amount()
print(f'Amount: {amount:,.4f}')
amount = ois.amount()
plazo = val_date.day_diff(ois.settlement_date)
df = zcc.get_discount_factor_at(plazo)
print(f'Df: {df:.6%}')
vp_check = df * amount
print(f'VP check es: {vp_check:,.4f}')

# Check tercer caso borde
val_date = Qcf.QCDate(1, 10, 2020)
set_expected_rate(val_date, ois, fixings, zcc)
pv = present_value(val_date, ois, zcc)
print(f'Valor presente es: {pv:,.4f}')
amount = ois.amount()
print(f'Amount: {amount:,.4f}')
amount = ois.amount()
plazo = val_date.day_diff(ois.settlement_date)
if plazo <= 0:
    df = 0.0
else:
    df = zcc.get_discount_factor_at(plazo)
print(f'Df: {df:.6%}')
vp_check = df * amount
print(f'VP check es: {vp_check:,.4f}')

# Check tercer caso
val_date = Qcf.QCDate(15, 10, 2020)
set_expected_rate(val_date, ois, fixings, zcc)
pv = present_value(val_date, ois, zcc)
print(f'Valor presente es: {pv:,.4f}')
amount = ois.amount()
print(f'Amount: {amount:,.4f}')
amount = ois.amount()
plazo = val_date.day_diff(ois.settlement_date)
if plazo <= 0:
    df = 0.0
else:
    df = zcc.get_discount_factor_at(plazo)
print(f'Df: {df:.6%}')
vp_check = df * amount
print(f'VP check es: {vp_check:,.4f}')
