==> 11_modelos_de_tasa_corta.ipynb <==
# Modelos de Tasa Corta

Comenzamos con algunas definiciones preliminares.

## Definiciones Preliminares

- Bank Account
- Factor de Descuento
- Tasa Compuesta Contínua
- Tasa Lineal
- Curva Cero Cupón
- Curva de Bonos Cero Cupón
- Tasa Forward Lineal
- Tasa Forward Instantánea

### Bank Account

Sea $B(t)$ el valor en $t\geq0$ de una cuenta bancaria que acumula intereses de forma contínua. Se asume que $B(0)=1$ y que la evolución en el tiempo de la cuenta está regida por la siguiente ecuación diferencial:

$$
dB(t)=r_tB(t)dt
$$

donde $r_t\geq 0$ $\forall t$. La solución de (1) es:

$$
B(t)=\exp{\left({\int_0}^t r_s ds\right)}
$$

La cantidad $r_t$ corresponde a la tasa instantánea que se aplica en el tiempo $t$.

Es evidente que si $r_t$ es una función determinística, el valor $B(t)$ es totalmente conocido en $t=0$. Por otra parte si $r_t$ es un proceso estocástico (una variable aleatoria $\forall t$ ), entonces $B(t)$ también lo es. Esta observación motiva la siguiente definición:

### Factor de Descuento

El factor de descuento (que puede ser estocástico) $D(t,T)$ entre los dos instante de tiempo $t$ y $T$ está dado por:

$$
D(t,T)=\frac{B(t)}{B(T)}=\exp\left(-{\int_t}^T r_s ds\right)
$$

**Contrariamente al modelamiento usual para Equity o FX, para instrumentos de tasa de interés siempre se supondrá que $r_t$ es un proceso estocástico.**

### Bono Cero Cupón

Un bono cero cupón que vence en $T$ (o bono a descuento) es un contrato que garantiza a su tenedor el pago de una unidad de moneda en el instante $T$. El valor de este contrato en $t\lt T$ se denota con $P(t,T)$. Claramente, $P(T,T)=1$ $\forall T$.

### Tasa Compuesta Contínua

La tasa compuesta contínua en $t$ para un instante de tiempo $T\gt t$ se denota con $R(t,T$) y se define a partir de la siguiente ecuación:

$$
P(t,T)=\exp\left[-R(t,T)\cdot (T-t)\right]
$$

Sigue que:

$$
R(t,T)=-\frac{\log P(t,T)}{T-t}
$$

### Tasa Lineal

De forma análoga, la tasa lineal se define a través de la siguiente ecuación:

$$
P(t,T)=\frac{1}{1+L(t,T)\cdot (T-t)}
$$

Y por lo tanto:

$$
L(t,T)=\left(\frac{1}{P(t,T)}-1\right)\cdot\frac{1}{T-t}
$$

### Curva Cero Cupón

Una curva cero cupón en $t$ es una función del tipo:

$$T\rightarrow\rho(t,T)$$

donde $\rho(t,T)$ representa una tasa en convención compuesta contínua, lineal o cualquier otra.

De forma similar se puede definir:

### Curva Bonos Cero Cupón

Una curva de bonos cero cupón en $t$ es una función del tipo:

$$T\rightarrow P(t,T)$$

donde $P(t,T)$ es el valor de.l bono cero cupón en $t$ que vence en $T$.

### Tasa Forward Lineal

La tasa forward lineal en $t$ que expira en $T$ y vence en $S$ con $S>T$ se denota con $F(t;T,S)$ y se define como:

$$
F(t;T,S):=\left(\frac{P(t,T)}{P(t,S)}-1\right)\cdot\frac{1}{S-T}=\left(\frac{W(t,S)}{W(t,T)}-1\right)\cdot\frac{1}{S-T}
$$

### Tasa Forward Instantánea

La tasa forward instantánea en $t$ que expira y vence en $T$ se denota con $f(t,T)$ y se define como:

$$
f(t,T):=\lim_{S \to T^+}F(t;T,S)=-\frac{\partial\log P(t,T)}{\partial T}
$$

Sigue que:

$$
P(t,T)=\exp\left(-{\int_t}^T f(t,u)\cdot du\right)
$$

## Ecuación Fundamental

- Aquí vemos la primera generación de modelos de tasa de interés en tiempo continuo. Los llamados modelos de la tasa corta.


- Se modela explícitamente la evolución de la tasa instantánea $r_t$ como única fuente de incertidumbre en tasa de interés.


- Esta tasa de interés se considera libre de riesgo financiero, es decir representa la tasa que se obtiene en un depósito a plazo (u otra fuente de financiamiento similar) entre $t$ y $t + dt$.


- De este modo, los valores de todos los instrumentos de tasa de interés quedan ligados a esta única variable estocástica.

Sea:

$$
dr_t=m\left(r_t,t\right)dt+\sigma\left(r_t,t\right)dX_t
$$

la ecuación diferencial estocástica en la medida histórica que gobierna la evolución de la tasa $r_t$. Aquí $X_t$ es un movimiento browniano en la medida histórica y las funciones $m\left(r_t,t\right)$ y $\sigma\left(r_t,t\right)$ son suficientemente regulares.

Queremos encontrar una ecuación a las derivadas parciales, al estilo de lo hecho para encontrar la fórmula de BSM, que nos permita valorizar instrumentos que dependen de la tasa de interés.

Sean $Z_1\left(r_t,t\right)$ y $Z_2\left(r_t,t\right)$ dos tales instrumentos y construyamos un portfolio con ambos que sea insensible a movimientos de tasa de interés, o sea:

$$
\Pi\left(r,t\right)=Z_1\left(r,t\right)-\Delta Z_2\left(r,t\right)
$$

$$
\frac{\partial\Pi\left(r,t\right)}{\partial r}=0
$$

$$
\Delta=\left(\frac{\partial Z_1}{\partial r}\right)\cdot\left(\frac{\partial Z_2}{\partial r}\right)^{-1}
$$

De ahora en adelante, por simplicidad, se omitirá la dependencia de $r$ de $t$.

Utilizando el lema de Itô tenemos que:

$$
d\Pi=\frac{\partial\Pi}{\partial t}dt+\frac{\partial\Pi}{\partial r}dr+\frac{1}{2}\frac{\partial^2\Pi}{\partial r^2}dr^2\\
=\left[\frac{\partial\Pi}{\partial t}+\frac{1}{2}\frac{\partial^2\Pi}{\partial r^2}\sigma^2\left(r,t\right)\right]dt+\frac{\partial\Pi}{\partial r}\left[m\left(r,t\right)dt+\sigma\left(r,t\right)dX_t\right]\\
=\left[\frac{\partial\Pi}{\partial t}+\frac{\partial\Pi}{\partial r}m\left(r,t\right)+\frac{1}{2}\frac{\partial^2\Pi}{\partial r^2}\sigma^2\left(r,t\right)\right]dt+\frac{\partial\Pi}{\partial r}\sigma\left(r,t\right)dX_t
$$

Dado que $\frac{\partial\Pi}{\partial r}=0$ tenemos que:

$$
d\Pi=\left(\frac{\partial\Pi}{\partial t}+\frac{1}{2}\frac{\partial^2\Pi}{\partial r^2}\sigma^2\left(r,t\right)\right)dt
$$

Luego, como el portfolio no es estocástico, por ausencia de arbitraje, sigue que:

$$
d\Pi=\Pi r dt
$$

y finalmente se obtiene que:

$$
\frac{\partial\Pi}{\partial t}+\frac{1}{2}\frac{\partial^2\Pi}{\partial r^2}\sigma^2\left(r,t\right)-r\Pi=0
$$

Recordando las definiciones de $\Pi$ y $\Delta$ llegamos a:

$$
\frac{\partial\Pi}{\partial t}=\frac{\partial Z_1}{\partial t}-\Delta\frac{\partial Z_2}{\partial t}\\
\frac{\partial^2\Pi}{\partial r^2}=\frac{\partial^2 Z_1}{\partial r^2}-\Delta\frac{\partial^2 Z_2}{\partial r^2}
$$

Sustituyendo y reordenando se llega a:

$$
\left(\frac{\partial Z_1}{\partial t}+\frac{1}{2}\frac{\partial^2 Z_1}{\partial r^2}\sigma^2\left(r,t\right)-r Z_1\right)\cdot\left(\frac{\partial Z_1}{\partial r}\right)^{-1}=\left(\frac{\partial Z_2}{\partial t}+\frac{1}{2}\frac{\partial^2 Z_2}{\partial r^2}\sigma^2\left(r,t\right)-r Z_2\right)\cdot\left(\frac{\partial Z_2}{\partial r}\right)^{-1}
$$

- Cambio de valor por paso del tiempo (*theta*):

$$
\frac{\partial Z}{\partial t}
$$


- Valor por convexidad o estocasticidad de las tasas de interés (*gamma*):

$$
\frac{1}{2}\frac{\partial^2 Z}{\partial r^2}
$$


- Costo de financiamiento de la posición (*carry*):

$$
rZ
$$


- Cambio de valor por cambio en las tasas de interés (*delta*):

$$
\frac{\partial Z}{\partial r}
$$

Se concluye que para todo instrumento $Z(r,t)$ vale la siguiente ecuación diferencial a las derivadas parciales:

$$
\left(\frac{\partial Z}{\partial t}+\frac{1}{2}\frac{\partial^2 Z}{\partial r^2}\sigma^2\left(r,t\right)-r Z\right)\cdot\left(\frac{\partial Z}{\partial r}\right)^{-1}=-m^*\left(r,t\right)\\
\frac{\partial Z}{\partial t}+\frac{\partial Z}{\partial r}m^*\left(r,t\right)+\frac{1}{2}\frac{\partial^2 Z}{\partial r^2}\sigma^2\left(r,t\right)-r Z=0
$$

Volviendo a hacer explícita la dependencia de $r$ de $t$, si el instrumento $Z(r,t)$ vence en $T$ podemos encontrar su valor $Z(r_t,t)$ en $t$ resolviendo la ecuación (27) junto con la oportuna condición terminal o de borde:

$$
Z(r_T, T)=g(r_T)
$$

Aquí la función $g(r_T)$ (que es una función del valor final $r_T$ de la tasa $r_t$) representa el payoff o valor final de $Z$.

Busquemos una interpretación para $m*$, definamos los siguientes dos objetos (partimos calculando la ecuación diferencial estocástica para un instrumento $Z$):

$$
dZ=\left(\frac{\partial Z}{\partial t}+\frac{1}{2}\frac{\partial^2 Z}{\partial r^2}\sigma^2\left(r,t\right)+\frac{\partial Z}{\partial r}m\left(r,t\right)\right)dt+\sigma\left(r,t\right)\frac{\partial Z}{\partial r}dX_t \\
\mu_Z\cdot Z=\frac{\partial Z}{\partial t}+\frac{1}{2}\frac{\partial^2 Z}{\partial r^2}\sigma^2\left(r,t\right)+\frac{\partial Z}{\partial r}m\left(r,t\right) \\
\sigma_Z\cdot Z=\sigma\left(r,t\right)\frac{\partial Z}{\partial r}
$$

Recordando (27) llegamos a:

$$
\frac{\mu_Z-r}{\sigma_Z}=\frac{m-m^*}{\sigma}:=\lambda
$$

Al cociente $\lambda$ se le llama *market price of interest rate risk*. Indica que la razón entre el exceso de retorno de un instrumento y su volatilidad es la misma para todos los instrumentos. Vemos además, el nexo que hay entre $m^*$ y $\lambda$.

En particular tenemos que $m*=m-\lambda\sigma$ y por lo tanto podemos escribir:

$$
\begin{equation}
\frac{\partial Z}{\partial t}+\frac{\partial Z}{\partial r}\left(m-\lambda\sigma\right)+\frac{1}{2}\frac{\partial^2 Z}{\partial r^2}\sigma^2\left(r,t\right)-r Z=0
\end{equation}
$$

## Fórmula de Feynman-Kac

Sea $V(r,t)$ el precio de un instrumento con payoff final igual a $V(r_T,T)=g(r_T,T)$ que satisfaga la siguiente ecuación diferencial parcial:

$$
\begin{equation}
\frac{\partial V}{\partial t}+\frac{\partial V}{\partial r}m^*\left(r_t,t\right)+\frac{1}{2}\frac{\partial^2 V}{\partial r^2}\sigma^2\left(r_t,t\right)=R\left(r_t\right)V
\end{equation}
$$

Donde $R(r)$ es una cierta función. Entonces $V(r_t,t)$ está dado por:

$$
\begin{equation}
V\left(r_t,t\right)=\mathbb{E}_t^*\left[e^{-\int_{t}^{T}R\left(r_u\right)du}g\left(r_T,T\right)\right] \\
\end{equation}
$$

Donde,

$$
\begin{equation}
dr_t=m^*\left(r_t,t\right)dt+\sigma\left(r_t,t\right)dX_t
\end{equation}
$$

## La Medida Forward

**Ref:** https://www.amazon.com/s?k=stochastic+finance+a+numeraire+approach&__mk_es_US=ÅMÅŽÕÑ&ref=nb_sb_noss

El cálculo de este valor esperado (fórmula de Feynman-Kac) es dificultoso porque la variable estocástica $r_t$ aparece en ambos términos del producto. Si recordamos que:

$$
\begin{equation}
\newcommand{\E}{\mathrm{E}}
\newcommand{\Var}{\mathrm{Var}}
\newcommand{\Cov}{\mathrm{Cov}}
\Cov\left(Z,Y\right)=\E\left(ZY\right)-\E\left(Z\right)\E\left(Y\right) \\
\E\left(ZY\right)=\E\left(Z\right)\E\left(Y\right)+\Cov\left(Z,Y\right)
\end{equation}
$$

Podemos escribir,

$$
\begin{equation}
\newcommand{\Cov}{\mathrm{Cov}}
\mathbb{E}_t^*\left[e^{-\int_{t}^{T}r_udu}g\left(r_T,T\right)\right]=\mathbb{E}_t^*\left[e^{-\int_{t}^{T}r_udu}\right]\mathbb{E}_t^*\left[g\left(r_T,T\right)\right]+\Cov\left(e^{-\int_{t}^{T}r_udu},g\left(r_T,T\right)\right)
\end{equation}
$$

Si pudiéramos eliminar el término con covarianza, quedaría algo mucho más sencillo, de hecho ya sabemos que:

$$
\begin{equation}
\newcommand{\Cov}{\mathrm{Cov}}
\mathbb{E}_t^*\left[e^{-\int_{t}^{T}r_udu}\right]=Z\left(t,T\right)
\end{equation}
$$

- Entendemos por numerario (del francés numeraire) la unidad en que expresamos los precios. Por ejemplo, en vez de usar USD, podríamos expresar precios de instrumentos financieros en términos del precio de un bono del tesoro prefijado.


- Veremos que un cambio de numerario lleva a una simplificación en las fórmulas de valoración de derivados.


- Sea $V(r_t,t,T)$ el valor en USD de un cierto instrumento y $Z(r_t,t,T)$ el valor del bono cupón cero que vence en T. Consideremos la siguiente renormalización o cambio de numerario:

$$
\begin{equation}
\overline{V}\left(r_t,t,T\right)=\frac{V\left(r_t,t,T\right)}{Z\left(r_t,t,T\right)}
\end{equation}
$$

El valor $\overline{V}\left(r_t,t,T\right)$ satisface la siguiente ecuación diferencial:

$$
\begin{equation}
\frac{\partial\overline{V}}{\partial t}+\frac{\partial\overline{V}}{\partial r}\left[m^*\left(r_t,t\right)+\frac{\partial Z}{\partial r}\frac{1}{Z}\sigma^2\left(r_t,t\right)\right]+\frac{1}{2}\frac{\partial^2\overline{V}}{\partial r^2}\sigma^2\left(r_t,t\right)=0
\end{equation}
$$

$$R(r_t)=0\quad \forall t$$


Si consideramos que:

$$
\begin{equation}
\frac{dZ}{Z}=\ldots+\sigma_{Z}dX_t
\end{equation}
$$

Tenemos que:

$$
\begin{equation}
\sigma_{Z}\left(r_t,t\right)=\frac{\partial Z}{\partial r}\frac{1}{Z}\sigma\left(r_t,t\right)
\end{equation}
$$

Y por lo tanto llegamos a:

$$
\begin{equation}
\frac{\partial\overline{V}}{\partial t}+\frac{\partial\overline{V}}{\partial r}\left[m^*\left(r_t,t\right)+\sigma_Z\left(r_t,t\right)\sigma\left(r_t,t\right)\right]+\frac{1}{2}\frac{\partial^2\overline{V}}{\partial r^2}\sigma^2\left(r_t,t\right)=0
\end{equation}
$$

Esta ecuación tiene dos diferencias importantes respecto a la ecuación fundamental: el término $rV$ desaparece y $m^*$ aparece modificado por el producto entre el coeficiente de difusión del bono cero cupón y el coeficiente de difusión de la tasa corta.

Podemos ahora aplicar la fórmula de Feynman-Kac para obtener que la solución a la ecuación diferencial anterior está dada por:

$$
\begin{equation}
\overline{V}\left(r_t,t,T\right)=\mathbb{E}_t^{Z_T}\left[g\left(r_T,T\right)\right]
\end{equation}
$$

Donde el valor esperado $\mathbb{E}_t^{Z_T}$ se calcula con la probabilidad tal que:

$$
\begin{equation}
dr_t=\left[m^*\left(r_t,t\right)+\sigma_Z\left(r_t,t\right)\sigma\left(r_t,t\right)\right]dt+\sigma\left(r_t,t\right)dX_t
\end{equation}
$$

El proceso anterior suele llamarse forward risk neutral o T-forward risk neutral cuando se desea enfatizar el plazo del instrumento a valorar. Finalmente tenemos que:

$$
\begin{equation}
V\left(r_t,t,T\right)=Z\left(r_t,t,T\right)\mathbb{E}_t^{Z_T}\left[g\left(r_T,T\right)\right]
\end{equation}
$$
