==> 15_ejemplo_cva.ipynb <==
# Ejemplo CVA

Se calcula el CVA de un swap.

## Librerías

- Las funciones relacionadas con el modelo de Hull-White que definimos y utilizamos en los notebooks 13 y 14 han sido mejoradas y encapsuladas en el módulo `hull_white.py`.

- La función `bono_tasa_fija` se trasladó al nuevo módulo `instruments.py`.

from modules import instruments as inst
from modules import hull_white as hw
from modules import auxiliary as aux
import pandas as pd
import numpy as np
import math

## Auxiliares

frmt = {'plazo': '{:,.0f}', 'tasa': '{:.8%}', 'df': '{:.6%}', 't': '{:.4f}'}

## Data Curva

Se lee la data y se completa con el plazo en convención Act/365.

df_curva = pd.read_excel('data/20201012_built_sofr_zero.xlsx')

df_curva['t'] = df_curva['plazo'] / 365.0

df_curva.tail().style.format(frmt)

## Funciones para el Modelo de HW

- `zrate`: es el cubic spline construido a partir de las columnas `t` y `tasa` de `df_curva`.
- `hwz`: es una versión más sencilla de la función `zero_hw` que sólo requiere llos parámetros `r`, `t` y `T`.
- `theta`: es la misma función que ya utilizamos.

gamma = 1.0
sigma = .0025

Se construyen las funciones `zrate`, `hwz` y `theta`.

zrate, hwz, theta = hw.get_zrate_hwzero_and_theta(
    df_curva['t'],
    df_curva['tasa'],
    gamma,
    sigma
)

Se verifica que `zrate` y más importante aún, `hwz` coincidan con los valores de la curva de mercado en los vértices de interpolación (plazos y tasas de la curva).

**Notar** el plazo utilizado para obtener el valor de `r0`.

r0 = zrate(.0000001)
print(f'         r0: {r0: .6%}\n')
for tup in df_curva.itertuples():
    print(f'      plazo: {tup.t:.4f}')
    print(f' tasa curva: {tup.tasa:.6%}')
    print(f'tasa spline: {zrate(tup.t):.6%}')
    print(f'   tasa hwz: {-math.log(hwz(r, 0, tup.t)) / tup.t:.6%}')
    print()

Se verifica que la función `theta` se mantenga acotada.

for tup in df_curva.itertuples():
    print(f'theta({tup.t:.4f}): {theta(tup.t):.6%}')

## Simulación

Se definen los parámetros requeridos, tanto del modelo como de la simulación.

num_sims = 1000
num_steps = 528 # 264 * 2 (2 años)

Se obtienen las trayectorias.

time_steps, paths = hw.sim_hw_many(gamma, sigma, theta, r0, 1000, 528, seed=1000)

**Ejercicio:** construir una función que retorne la función `sim_hw_many` pero sólo con los argumentos r0, num_sims, num_:steps y seed. O sea, que gamma, sigma y theta queden encerrados (closure) dentro de la función.

time_steps[0:20]

paths[0][0:10] # son los primeros 10 valores de la primera trayectoria simulada.

Se extrae el elemento correspondiente a 6M, 1Y, 18M de cada simulación:

EE = Promedio(max(Vi, 0)) y se va a calcular en estos tiempos de parada o stopping times.

last_r = [(s[132], s[264], s[396]) for s in paths]
last_r[0]

## Cálculo de la *EE* de un Swap

Para el swap vamos a utilizar una función sencilla para la pata fija y consideraremos que el valor presente de la pata flotante siempre es igual al nocional. Consideremos un swap a 2Y.

swap_fija = inst.bono_tasa_fija(0, .5, 4, .001)

swap_fija

Se define una función que construye una función que valoriza el swap a partir de $t$ y $r_t$.

def valorizador_swap(pata_fija, df_function):
    def valor(r, t):
        result = 0.0
        for e in pata_fija:
            if e[0] > t:
                result += e[1] * df_function(r, t, e[0])
        return result - 100.0 if result > 0.0 else 0.0
    return valor

Probemos:

vswap = valorizador_swap(swap_fija, hwz)

vswap(r0, 2)

Valoricemos el swap en 6M más en cada simulación:

valor_swap_6m = [vswap(r[0], .5) for r in last_r]
valor_swap_6m[0: 10]

Mejor aún, podemos calcular directamente la exposición.

e_swap_6m = [max(vswap(r[0], .5), 0) for r in last_r]
e_swap_6m[0: 10]

Por lo tanto, la exposición esperada en 6M es:

ee_swap_6m = np.average(e_swap_6m)

print(f'La EE del swap en 6M es: {ee_swap_6m: ,.2f}')

**Ejercicio:** completar el cálculo del CVA.

Se calcula la *EE* a 1Y y a 18M.

e_swap_1y = [max(vswap(r[1], 1), 0) for r in last_r]
e_swap_1y[0: 10]

e_swap_18m = [max(vswap(r[2], 1.5), 0) for r in last_r]
e_swap_18m[0: 10]

Luego, se calcula el factor de descuento a cada uno de los plazos en que se calcula *EE*: 6M, 1Y y 18M.

- Se extraen los resultados de la simulación hasta 6M, 1Y y 18M.

paths_up_to_6m = [s[0:132] for s in paths]
paths_up_to_1y = [s[0:264] for s in paths]
paths_up_to_18m = [s[0:396] for s in paths]

Luego, se calculan los factores de descuento "devolviéndose por cada trayectoria".

dt = 1 / 264.0

dfs_up_to_6m = [np.exp(-dt * np.sum(path)) for path in paths_up_to_6m]
dfs_up_to_1y = [np.exp(-dt * np.sum(path)) for path in paths_up_to_1y]
dfs_up_to_18m = [np.exp(-dt * np.sum(path)) for path in paths_up_to_18m]

pv_ee_6m = [both[0] * both[1] for both in zip(e_swap_6m, dfs_up_to_6m)]
pv_ee_1y = [both[0] * both[1] for both in zip(e_swap_1y, dfs_up_to_1y)]
pv_ee_18m = [both[0] * both[1] for both in zip(e_swap_18m, dfs_up_to_18m)]

mean_pv_ee_6m = np.mean(pv_ee_6m)
mean_pv_ee_1y = np.mean(pv_ee_1y)
mean_pv_ee_18m = np.mean(pv_ee_18m)
print(mean_pv_ee_6m, mean_pv_ee_1y, mean_pv_ee_18m)
