==> 17_pauta_tarea_2.ipynb <==
# Pauta Tarea 2

- Usando las funciones anteriores, o desarrollando las propias, escriba una función que pueda valorizar, por simulación de Montecarlo, calls y puts sobre bonos cero cupón con el modelo de HW. Considere puts y calls de plazo máximo igual a 1Y.
- Su función debe poder recibir la semilla.
- Compare sus resultados con las fórmulas del modelo y muestre que diferencias obtiene usando 1000 simulaciones y una semilla predeterminada.
- Valorice también por simulación de Montecarlo el bono cupón cero a 1Y. Muestre que diferencias obtiene respecto al valor de mercado del bono usando 1000 simulaciones y una semilla predeterminada (valor de mercado significa el valor que hoy se puede observar .

## Configuración

### Librerías

from modules import hull_white as hw
from modules import auxiliary as aux
from typing import Callable
import pandas as pd
import numpy as np
import math

### Formatos para DataFrames

frmt = {'plazo': '{:,.0f}', 'tasa': '{:.8%}', 'df': '{:.6%}', 't': '{:.4f}',
        'analytic_put': '{:.6%}', 'sim_put': '{:.6%}', 'dif_put': '{:.6%}',
        'analytic_call': '{:.6%}', 'sim_call':  '{:.6%}', 'dif_call': '{:.6%}',
        'analytic_zero': '{:.6%}', 'sim_zero':  '{:.6%}', 'dif_zero': '{:.6%}', }

### Constante `DT`

DT = 1 / 264.0

### Curva Cero Cupón

df_curva = pd.read_excel('data/20201012_built_sofr_zero.xlsx')
df_curva['t'] = df_curva['plazo'] / 365.0

df_curva.tail().style.format(frmt)

### Funciones del Modelo HW

- `zrate`: es el cubic spline construido a partir de las columnas `t` y `tasa` de `df_curva`.
- `hwz`: es una versión más sencilla de la función `zero_hw` que sólo requiere llos parámetros `r`, `t` y `T`.
- `theta`: es la misma función que ya utilizamos.

Primero se establecen los valores de $\gamma$ y $\sigma$.

gamma = 1
sigma = .005

Se construyen las funciones `zrate`, `hwz` y `theta`.

zrate, hwz, theta = hw.get_zrate_hwzero_and_theta(
    df_curva['t'],
    df_curva['tasa'],
    gamma,
    sigma
)

## Función para Valorizar

Primero se define la función `payoff` para las call y put sobre un bono cupón cero.

def call_put_payoff(
    c_p: hw.CallPut,
    strike: float,
    t: float,
    T: float,
    r: float,
    hwz: Callable[[float, float, float], float]) -> float:
    """
    """
    if c_p == hw.CallPut.CALL:
        return max(hwz(r, t, T) - strike, 0.0)
    else:
        return max(strike - hwz(r, t, T), 0.0)

Se fijan los parámetros `c_p`, `strike`, `t`, `T` y `hwz`.

def get_call_put_payoff(
    c_p: hw.CallPut,
    strike: float,
    t: float,
    T: float,
    hwz: Callable[[float, float, float], float]) -> Callable[[float,], float]:
    """
    """
    def payoff(r: float):
        return call_put_payoff(c_p, strike, t, T, r, hwz)
    return payoff

Se fijan los parámetros de simulación de la función `hw.sim_hw_many`.

def get_simulator(
    gamma: float,
    sigma: float,
    theta: Callable[[float,], float],
    num_sim: int,
    num_steps: int,
    r0: float,
    seed=None):
    """
    """
    def simulator(seed=None):
        return hw.sim_hw_many(gamma, sigma, theta, r0, num_sim, num_steps, seed)
    return simulator

Se definen `payoff` y `simulator`.

call = get_call_put_payoff(hw.CallPut.CALL, .99, 1, 2, hwz)

simulator = get_simulator(gamma, sigma, theta, 100, 264, zrate(.000001), 16127)

Finalmente, con estos últimos dos elementos se define la función de valorización:

def value_payoff_sim(simulation, payoff) -> float:
    """
    """
    result = 0.0
    for s in simulation:
        result += math.exp(-DT * np.sum(s)) * payoff(s[-1])
    return result / len(simulation)

time, simulation = simulator()

value = value_payoff_sim(simulation, call)
print(f'simulated value: {value:.6%}')

## Diferencias con Calls y Puts

- Se muestra las diferencias para Calls y Puts y strikes desde 98% hasta 102% en intervalos de 0.5%.
- Los subyacentes son bonos cupón cero de 2Y, 3Y, 4Y y 5Y.
- Todas las opciones tienen vencimiento en 1Y.

compare = []
simulator = get_simulator(gamma, sigma, theta, 100, 264, zrate(.000001), 16127)
time, simulation = simulator()
for T in [2, 3, 4, 5]:
    for strike in [.9800, .9850, .9900, .1000, .1005, .1010, .1015, .1020]:
        zo = math.exp(-zrate(1))
        zb = math.exp(-zrate(T) * T)
        
        analytic_call = hw.zcb_call_put(hw.CallPut.CALL, strike, 1, T, zrate(.000001), zo, zb, gamma, sigma)
        call = get_call_put_payoff(hw.CallPut.CALL, strike, 1, T, hwz)
        sim_call = value_payoff_sim(simulation, call)
        
        analytic_put = hw.zcb_call_put(hw.CallPut.PUT, strike, 1, T, zrate(.000001), zo, zb, gamma, sigma)
        put = get_call_put_payoff(hw.CallPut.PUT, strike, 1, T, hwz)
        sim_put = value_payoff_sim(simulation, put)
        
        compare.append((T, strike,
                        analytic_call, sim_call, analytic_call - sim_call,
                        analytic_put, sim_put, analytic_put - sim_put))

df_compare = pd.DataFrame(compare)
df_compare.columns = ['T', 'strike',
                      'analytic_call', 'sim_call', 'dif_call',
                      'analytic_put', 'sim_put', 'dif_put']

df_compare.style.format(frmt)

## Diferencias con Bonos Cupón Cero

def zcb(r: float):
    return 1

simulator = get_simulator(gamma, sigma, theta, 1000, 20 * 264, zrate(.000001), 16127)
time, simulation = simulator()

compare2 = []
for T in [1, 2, 3, 4, 5, 10, 12, 15, 20]:
    analytic_zero = math.exp(-zrate(T) * T)
    stop_index = list(time).index(T)
    simulazione = simulation[:, 0:stop_index]
    sim_zero = value_payoff_sim(simulazione, zcb)

    compare2.append((T, analytic_zero, sim_zero, analytic_zero - sim_zero))

df_compare2 = pd.DataFrame(compare2)
df_compare2.columns = ['T', 'analytic_zero', 'sim_zero', 'dif_zero']

df_compare2.style.format(frmt)
